// src/api/orval/model/notificationRequest.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */
import type { NotificationSeverity } from "./notificationSeverity";

export interface NotificationRequest {
  severity?: NotificationSeverity;
  /** @nullable */
  serverId?: number | null;
  /** @nullable */
  actorUserId?: number | null;
  /** @nullable */
  relatedClientId?: number | null;
  /** @nullable */
  correlationId?: string | null;
  /** @nullable */
  dedupKey?: string | null;
  /** @nullable */
  type?: string | null;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  message?: string | null;
  /** @nullable */
  source?: string | null;
}