// src/api/orval/model/notificationItemDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * DataGate Monitor API
 * OpenAPI spec version: v1
 */
import type { NotificationSeverity } from "./notificationSeverity";

export interface NotificationItemDto {
  id?: number;
  /** @nullable */
  type?: string | null;
  severity?: NotificationSeverity;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  message?: string | null;
  isRead?: boolean;
  createdAt?: string;
  /** @nullable */
  readAt?: string | null;
}