// src/api/orval/model/getNotificationsResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */
import type { NotificationItemDtoPagedResponse } from "./notificationItemDtoPagedResponse";

export interface GetNotificationsResponse {
  notifications?: NotificationItemDtoPagedResponse;
}